ElectricCircuit: do not skip valid circuits for other nodes changing state
Previously we did not have de-duplication code for open
circuits and so we tried this tricks to prevent registering
a duplicate circuit in advance so we did not have to
check later for duplicates.

This tricks were not working so now we register all and
later check for duplicates and remove them.

A node like an on/off contact can have multiple open
circuits waiting to be extended when it goes on.

If this node is after a parallel and on negative side
(second pole, now possible with BifilarizatorNode)
then multiple open circuits will flow to same exit contact.

Try all of them otherwise we skip valid circuits.
If some are duplicates it will be taken care of later!